dok/requirements.txt
absl-py==2.1.0
altair==5.4.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.4.2
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
fqdn==1.5.1
fsspec==2024.10.0
future==0.18.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.0
gunicorn==20.0.4
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.6.4
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
lab==6.2
lazy-object-proxy==1.4.3
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.14.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.2.4
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.4.10
platformdirs==4.3.6
preprocessing==0.1.13
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.6.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pybboxes==0.1.6
pycodestyle==2.6.0
pycparser==2.22
pydeck==0.9.1
pyflakes==2.2.0
Pygments==2.18.0
pylint==2.6.0
pyparsing==3.2.0
pyrsistent==0.17.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.0.1
QtPy==1.9.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
sahi==0.11.19
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.19.0
setuptools==75.5.0
shapely==2.0.6
simplejson==3.17.2
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sphinx-rtd-theme==0.2.4
stack-data==0.6.3
streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
testpath==0.4.4
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
txt2tags==3.7
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.43
ultralytics-thop==2.0.11
ultralyticsplus==0.0.28
uri-template==1.3.0
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==1.2.1
